Fix missing dependencies for docker
Previously, it was failing to start the docker container and was
complaining about concurrent-ruby dependency. This was because
Gemfile.lock was pointing to a version of the dependency that doesn't
exist anymore.

This commit upgrades all the dependencies and adds some debugging logs
so that you can see when the application has started.