Upgrade testmine to use ruby 2.5.1 and rails 5.0
Github reported several vulnerabilities with the gems testmine was
using. It also was using an unsupported Ruby version i.e 1.9.

This commit upgrades ruby to 2.5 and rails to 5.0, rest are the
side-effects of these upgrades.

Also verified on Docker that it works fine and there were no new
reported errors in the logs.